+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 - 2023-01-22
 
 - Remove `jest-mock-console` dependency
+- Initialize `repository`, `homepage`, `bugs` and `author` fields in `package.json` from Git data
 
 ## 1.0.0 - 2022-12-17